% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAST_UPDATE}
\alias{LAST_UPDATE}
\title{The date at which the C-E datasets were last updated}
\format{
An object of class \code{Date} of length 1.
}
\usage{
LAST_UPDATE
}
\description{
The date at which the C-E datasets were last updated
}
\keyword{datasets}
