% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METADATA}
\alias{METADATA}
\title{A list of metadata fields describing the various datasets.
These include the number of rows for each dataset and their date of last update}
\format{
An object of class \code{list} of length 4.
}
\usage{
METADATA
}
\description{
A list of metadata fields describing the various datasets.
These include the number of rows for each dataset and their date of last update
}
\keyword{datasets}
