% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAW.EF}
\alias{RAW.EF}
\title{The 'raw' monthly effort dataset}
\format{
\describe{
  \item{YEAR}{}
  \item{QUARTER}{}
  \item{MONTH_START}{}
  \item{MONTH_END}{}
  \item{FISHING_GROUND_CODE}{}
  \item{FISHING_GROUND}{}
  \item{FLEET_CODE}{}
  \item{FLEET}{}
  \item{FISHERY_TYPE_CODE}{}
  \item{FISHERY_TYPE}{}
  \item{FISHERY_GROUP_CODE}{}
  \item{FISHERY_GROUP}{}
  \item{FISHERY_CODE}{}
  \item{FISHERY}{}
  \item{GEAR_CODE}{}
  \item{GEAR}{}
  \item{SCHOOL_TYPE_CODE}{}
  \item{EFFORT}{}
  \item{EFFORT_UNIT_CODE}{}
}
}
\usage{
RAW.EF
}
\description{
The 'raw' monthly effort dataset
}
\keyword{datasets}
