% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RSD.CA}
\alias{RSD.CA}
\title{The 'raised' monthly catch dataset in weight and number}
\format{
\describe{
  \item{YEAR}{}
  \item{QUARTER}{}
  \item{MONTH_START}{}
  \item{MONTH_END}{}
  \item{FISHING_GROUND_CODE}{}
  \item{FISHING_GROUND}{}
  \item{FLEET_CODE}{}
  \item{FLEET}{}
  \item{FISHERY_TYPE_CODE}{}
  \item{FISHERY_TYPE}{}
  \item{FISHERY_GROUP_CODE}{}
  \item{FISHERY_GROUP}{}
  \item{FISHERY_CODE}{}
  \item{FISHERY}{}
  \item{GEAR_CODE}{}
  \item{GEAR}{}
  \item{CATCH_SCHOOL_TYPE_CODE}{}
  \item{IUCN_STATUS_CODE}{}
  \item{IUCN_STATUS}{}
  \item{SPECIES_WP_CODE}{}
  \item{SPECIES_WP}{}
  \item{SPECIES_GROUP_CODE}{}
  \item{SPECIES_GROUP}{}
  \item{SPECIES_CATEGORY_CODE}{}
  \item{SPECIES_CATEGORY}{}
  \item{SPECIES_CODE}{}
  \item{SPECIES}{}
  \item{SPECIES_SCIENTIFIC}{}
  \item{SPECIES_FAMILY}{}
  \item{SPECIES_ORDER}{}
  \item{IS_IOTC_SPECIES}{}
  \item{IS_SPECIES_AGGREGATE}{}
  \item{IS_SSI}{}
  \item{CATCH}{}
  \item{CATCH_UNIT_CODE}{}
  \item{CATCH_IN_NUMBERS}{}
}
}
\usage{
RSD.CA
}
\description{
The 'raised' monthly catch dataset in weight and number
}
\keyword{datasets}
